+++ docs/conf.py
@@ -27,13 +27,7 @@ def no_namedtuple_attrib_docstring(app, what, name,
         del lines[:]
 
 
-def skip(app, what, name, obj, would_skip, options):
-    if name == "__init__":
-        return False
-    return would_skip
-
 def setup(app):
-    app.connect("autodoc-skip-member", skip)
     app.connect(
         'autodoc-process-docstring',
         no_namedtuple_attrib_docstring,